pyproject.toml
[tool.poetry]
name = "powermeter"
version = "0.1.0"
description = ""
authors = ["Gilles Lescot <lescotg@hotmail.com>"]

[tool.poetry.dependencies]
python = "3.5.3"
flask = "*"
authlib = "*"
python-dotenv = "^0.10.1"

[tool.poetry.dev-dependencies]
pylint = "^2.3"
